Refactored utils for explorer and client
