+++ heron/cli/src/python/execute.py
@@ -17,10 +17,12 @@ import os
 import subprocess
 import tarfile
 import tempfile
+import traceback
 
 import heron.cli.src.python.opts  as opts
-import heron.common.src.python.utils as utils
+import heron.common.src.python.utils.config as utils
 import heron.cli.src.python.jars  as jars
+import heron.common.src.python.pex_loader as pex_loader
 
 
 ################################################################################
@@ -31,7 +33,7 @@ def heron_class(class_name, lib_jars, extra_jars=None, args=None, java_defines=N
   :param lib_jars:
   :param extra_jars:
   :param args:
-  :param javaDefines:
+  :param java_defines:
   :return:
   '''
   # default optional params to empty list if not provided
@@ -99,3 +101,16 @@ def heron_tar(class_name, topology_tar, arguments, tmpdir_root, java_defines):
 
   # Now execute the class
   heron_class(class_name, lib_jars, extra_jars, arguments, java_defines)
+
+def heron_pex(topology_pex, topology_class_name, tmp_dir):
+  if opts.verbose():
+    print("Importing " + topology_class_name + " from " + topology_pex)
+  try:
+    pex_loader.load_pex(topology_pex)
+    topology_class = pex_loader.import_and_get_class(topology_pex, topology_class_name)
+    topology_class.write(tmp_dir)
+  except Exception as e:
+    traceback.print_exc()
+    err_str = "Topology pex failed to be loaded. Bailing out..."
+    raise RuntimeError(err_str)
+
