+++ heron/cli/src/python/restart.py
@@ -17,7 +17,7 @@ from heron.common.src.python.color import Log
 import heron.cli.src.python.args as args
 import heron.cli.src.python.execute as execute
 import heron.cli.src.python.jars as jars
-import heron.common.src.python.utils as utils
+import heron.common.src.python.utils.config as utils
 
 
 def create_parser(subparsers):
