+++ heron/cli/src/python/submit.py
@@ -24,7 +24,7 @@ import heron.cli.src.python.args as cli_args
 import heron.cli.src.python.execute as execute
 import heron.cli.src.python.jars as jars
 import heron.cli.src.python.opts as opts
-import heron.common.src.python.utils as utils
+import heron.common.src.python.utils.config as utils
 
 
 ################################################################################
@@ -89,7 +89,7 @@ def launch_a_topology(cl_args, tmp_dir, topology_file, topology_defn_file):
       "--release_file", release_yaml_file,
       "--topology_package", topology_pkg_path,
       "--topology_defn", topology_defn_file,
-      "--topology_jar", topology_file
+      "--topology_jar", topology_file   # pex file if pex specified
   ]
 
   if opts.verbose():
@@ -235,6 +235,31 @@ def submit_tar(cl_args, unknown_args, tmp_dir):
 
   return True
 
+################################################################################
+#  Execute the pex file to create topology definition file by running
+#  the topology's main class.
+################################################################################
+def submit_pex(cl_args, unknown_args, tmp_dir):
+  # execute main of the topology to create the topology definition
+  # TODO: currently topology's class name is not necessary
+  topology_file = cl_args['topology-file-name']
+  topology_class_name = cl_args['topology-class-name']
+  try:
+    execute.heron_pex(topology_file, topology_class_name, tmp_dir)
+  except Exception as ex:
+    Log.error("Unable to execute topology main class: " + ex.message)
+    return False
+
+  try:
+    launch_topologies(cl_args, topology_file, tmp_dir)
+
+  except Exception as ex:
+    return False
+
+  finally:
+    shutil.rmtree(tmp_dir)
+
+  return True
 
 ################################################################################
 # pylint: disable=unused-argument
@@ -257,14 +282,15 @@ def run(command, parser, cl_args, unknown_args):
 
   # check to see if the topology file exists
   if not os.path.isfile(topology_file):
-    Log.error("Topology jar|tar file %s does not exist" % topology_file)
+    Log.error("Topology jar|tar|pex file %s does not exist" % topology_file)
     return False
 
   # check if it is a valid file type
   jar_type = topology_file.endswith(".jar")
   tar_type = topology_file.endswith(".tar") or topology_file.endswith(".tar.gz")
-  if not jar_type and not tar_type:
-    Log.error("Unknown file type. Please use .tar or .tar.gz or .jar file")
+  pex_type = topology_file.endswith(".pex")
+  if not jar_type and not tar_type and not pex_type:
+    Log.error("Unknown file type. Please use .tar or .tar.gz or .jar or .pex file")
     return False
 
   # create a temporary directory for topology definition file
@@ -286,5 +312,8 @@ def run(command, parser, cl_args, unknown_args):
 
   elif tar_type:
     return submit_tar(cl_args, unknown_args, tmp_dir)
+
+  elif pex_type:
+    return submit_pex(cl_args, unknown_args, tmp_dir)
 
   return False
