+++ heron/explorer/src/python/args.py
@@ -13,7 +13,7 @@
 # limitations under the License.
 ''' args.py '''
 import os
-import heron.common.src.python.utils as utils
+import heron.common.src.python.utils.config as utils
 
 # default parameter - url to connect to heron tracker
 DEFAULT_TRACKER_URL = "http://localhost:8888"
