+++ heron/explorer/src/python/clusters.py
@@ -15,7 +15,7 @@
 import heron.explorer.src.python.args as args
 # from heron.common.src.python.color import Log
 # from tabulate import tabulate
-import heron.common.src.python.utils as utils
+import heron.common.src.python.utils.tracker_access as tracker_access
 from heron.common.src.python.color import Log
 
 
@@ -35,7 +35,7 @@ def create_parser(subparsers):
 def run(command, parser, cl_args, unknown_args):
   """ run command """
   try:
-    clusters = utils.get_clusters()
+    clusters = tracker_access.get_clusters()
   except:
     Log.error("Fail to connect to tracker: \'%s\'", cl_args["tracker_url"])
     return False
